Split HNSW Kani proofs by concern
Replace the monolithic HNSW Kani proof file with a module directory that
keeps shared helpers in `mod.rs` and groups bidirectional, eviction, and
invariant harnesses in focused files below the 400-line limit.

Define `CBMC` on first use in the hypothesis document and update live
verification documentation to reference the new HNSW proof module path.
